requirements.txt
asgiref==3.5.2
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.1.3
django-allauth==0.54.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-jsonfield==1.4.1
django-location-field==2.7.1
django-oauth-toolkit==2.3.0
django-rest-api==0.1.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-stripe-payments==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.3.0
drf-yasg==1.21.6
google-auth==2.20.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
paystack==1.5.0
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
stripe==4.2.0
typing_extensions==4.6.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.5.0
wrapt==1.15.0
